creado el form component con modificacion de detalle de diseño
