+++ package.json
@@ -14,8 +14,9 @@
     "next": "15.3.5",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-icons": "^5.5.0",
     "react-toastify": "^11.0.5",
-    "sweetalert2": "^11.22.2",
+    "sweetalert2": "^11.6.13",
     "yup": "^1.6.1"
   },
   "devDependencies": {
